Add some notes, full re-run, use cabin number, standardize all features
The code also now uses nearly all the features.  A note is left about
what features were present beforehand, since this setup seems to
perform similarly with random forests and SVC.  It may help with
xgboost, though.

Features are now standardized as well, which seems to help performane
of SVM and possibly xgboost.